+++ newrelic-security-agent/src/main/java/com/newrelic/agent/security/intcodeagent/logging/HealthCheckScheduleThread.java
@@ -100,7 +100,7 @@ public class HealthCheckScheduleThread {
     private HealthCheckScheduleThread() {}
 
     public void scheduleNewTask() {
-        future = SchedulerHelper.getInstance().scheduleHealthCheck(runnable, 1, 5, TimeUnit.MINUTES);
+        future = SchedulerHelper.getInstance().scheduleHealthCheck(runnable, 30, 300, TimeUnit.SECONDS);
     }
 
     public boolean cancelTask(boolean forceCancel) {